// types.ts
export interface TaskTemplate {
    id: string;
    plan_code: string;
    asset_category: string;
    title: string;
    task_type: string;
    criticality: string;
    classification1?: string;
    classification2?: string;
    estimated_duration_minutes: number;
    planned_downtime_minutes?: number;
    frequency: string;
    frequency_days: number;
    subtasks: any[]; 
}